s3-rotate-cdk/lambda/requirements.txt
Pillow==10.4.0
boto3==1.34.160
